+++ R/largeVis.R
@@ -17,6 +17,7 @@
 #' the first pass through the neighborhood exploration phases will involve up to \eqn{N * nTrees * threshold} comparisons.
 #' @param max_depth See \code{\link{randomProjectionTreeSearch}}
 #' @param max_iter See \code{\link{randomProjectionTreeSearch}}.
+#' @param distance_method One of "Euclidean" or "Cosine."  See \code{\link{randomProjectionTreeSearch}}.
 #' @param perplexity See paper
 #' @param sgd_batches See \code{\link{projectKNNs}}.
 #' @param M See \code{\link{projectKNNs}}.
@@ -30,7 +31,7 @@
 #' @param ... See paper
 #'
 #' @return A `largeVis` object with the following slots:
-#'  \itemize{
+#'  \describe{
 #'    \item{'knns'} {An [N,K] integer matrix, which is an adjacency list of each vertex' identified nearest neighbors.
 #'    If the algorithm failed to find \code{K} neighbors, the matrix is padded with \code{NA}'s.}
 #'    \item{'wij'} {A sparse [N,N] matrix where each cell represents \eqn{w_{ij}}.}
@@ -76,13 +77,14 @@ vis <- function(x,
                      tree_threshold = max(10, nrow(x)),
                      max_iter = 3,
                      max_depth = 32,
+                     distance_method = "Euclidean",
 
                      perplexity = 50,
 
                      sgd_batches = ncol(x) * 20000,
                      M = 5,
                      weight_pos_samples = TRUE,
-                     alpha = 2,
+                     alpha = 1,
                      gamma = 7,
                      rho = 1,
                      min_rho = 0,
@@ -94,7 +96,7 @@ vis <- function(x,
   N <- ncol(x)
 
   if (check.assumptions)   {
-    if ((any(is.na(x)) +
+    if ( (any(is.na(x)) +
          any(is.infinite(x)) +
          any(is.nan(x))) > 0)
       stop("Missing values present in input matrix.")
@@ -109,6 +111,7 @@ vis <- function(x,
                                      K = K,
                                      max_iter = max_iter,
                                      max_depth = max_depth,
+                                     distance_method = distance_method,
                                      verbose = verbose)
 
   #############################################
@@ -118,7 +121,7 @@ vis <- function(x,
   # These vectors are analogous to the components of a sparse matrix,
   # but both triple and C-compressed forms are created.
   # The i and j vectors are 0-indexed while p is 1-indexed.
-  is <- rep(0:(N-1), each = K)
+  is <- rep(0:(N - 1), each = K)
   js <- as.vector(knns)
   is <- is[! js == -1]
   js <- js[! js == -1]
@@ -128,70 +131,39 @@ vis <- function(x,
   ord <- order(is)
   is <- is[ord]
   js <- js[ord]
-  ps <- rep(NA,N + 1)
-  diffs <- diff(is)
-  ps[is[which(diffs > 0)] + 2] <- which(diffs > 0) + 1
-  good <- cumsum(!is.na(ps))
-  ps <- ps[good + 1]
-  ps[1] <- 1
-  ps[length(ps) + 1] <- length(is)
 
   #######################################################
   # Calculate edge weights for candidate neighbors
   #######################################################
   if (verbose) cat("Calculating neighbor distances.\n")
 
-  xs <- distance(is, js, x, verbose)[,1]
+  xs <- distance(is, js, x, distance_method,verbose)[, 1]
 
   if (verbose) cat("\n")
 
-  if ((any(is.na(xs)) + any(is.infinite(xs)) + any(is.nan(xs)) + any(xs == 0)) > 0)
+  if ( (any(is.na(xs)) + any(is.infinite(xs)) + any(is.nan(xs)) + any(xs == 0)) > 0)
   stop("An error leaked into the distance calculation - check for duplicates")
-
-  ########################################################
-  # Estimate sigmas
-  ########################################################
-  if (verbose) {
-    progress = txtProgressBar(max = N, title = "sigmas")
-    cat("Estimating sigmas\n")
-  }
-
-  perplexity = log2(perplexity)
-  sigmas <- parallel::mclapply(1:N, FUN = function(idx) {
-    if (verbose) setTxtProgressBar(progress, idx)
-    x_i <- xs[ps[idx]:(ps[idx + 1] - 1)]
-    ret <- optimize(f = sigFunc,
-             x = x_i,
-             perplexity = perplexity,
-             interval = c(0,10000))
-  })
-  sigmas <- sapply(sigmas, `[[`, 1)
-
-  if (verbose) close(progress)
-
-  if (any(is.na(sigmas)) + any(is.infinite(sigmas)) + any(is.nan(sigmas)) + any((sigmas == 0)) > 0)
-    stop("An error has propogated into the sigma vector.")
+  if (any(xs > 27)) stop(paste(
+    "The Distances between some neighbors are large enough to cause the calculation of p_{j|i} to overflow.",
+    "Consider scaling the data matrix or using an alternative distance function."))
 
   #######################################################
-  # Calculate w_{ij}
+  # Get w_{ij}
   #######################################################
 
-  if (! requireNamespace('Matrix',quietly=T)) stop("The Matrix package must be available.")
+  ps <- i2p(is)
+  sigwij <- buildEdgeMatrix(i = is,
+                         j = js,
+                         p = ps,
+                         d = xs,
+                         perplexity = perplexity,
+                         verbose = verbose)
 
-  if (verbose) cat("Calculating w_{ij}.\n")
-  wij <- distMatrixTowij(is, js, xs, sigmas, N, verbose)
-
-  if (any(is.na(wij@x)) || any(is.infinite(wij@x)) || any(is.nan(wij@x)) || any((wij@x == 0)) > 0)
-    stop("An error has propogated into the w_{ij} vector.  This probably means the input data wasn't scaled.")
-
-  # Symmetricize
-  wij <- wij + Matrix::t(wij)
-  rm(xs, js, is)
 
   #######################################################
   # Estimate embeddings
   #######################################################
-  coords <- projectKNNs(wij = wij,
+  coords <- projectKNNs(wij = sigwij$wij,
                         dim = dim,
                         sgd_batches = sgd_batches,
                         M = M,
@@ -211,13 +183,13 @@ vis <- function(x,
 
   returnvalue <- list(
     knns = t(knns),
-    wij = wij,
+    wij = sigwij$wij,
     call = sys.call(),
     coords = coords,
-    sigmas = sqrt(sigmas / 2)
+    sigmas = sqrt(sigwij$sigmas / 2)
   )
 
-  class(returnvalue) <- 'largeVis'
+  class(returnvalue) <- "largeVis"
   return(returnvalue)
 }