+++ R/projectKNNs.R
@@ -10,13 +10,17 @@
 #'
 #' The objective function is: \deqn{ O = \sum_{(i,j)\in E} w_{ij} (\log f(||p(e_{ij} = 1||) + \sum_{k=1}^{M} E_{jk~P_{n}(j)} \gamma \log(1 - f(||p(e_{ij_k} - 1||)))}
 #' where \eqn{f()} is a probabilistic function relating the distance between two points in the low-dimensional projection space,
-#' and the probability that they are nearest neighbors.  See the discussion of the alpha parameter below.
+#' and the probability that they are nearest neighbors.
+#'
+#' There are two available probabilistic functions, \eqn{1 / (1 + \alpha \dot ||x||^2)} and \eqn{1 / (1 + \exp(||x||^2))}.
+#' The second function, which the paper authors recommend against, is used if parameter \code{alpha} is set to 0.
 #'
 #' The \code{weight_pos_samples} parameter controls how to handle edge-weights.  The paper authors recommend using a weighted
 #' sampling approach to select edges, and treating edge-weight as binary in calculating the objective. This is the default.
 #'
 #' However, the algorithm for drawing weighted samples runs in \eqn{O(n \log n)}. The alternative approach, which runs in
-#' \eqn{O(n)}, is to draw unweighted samples and include \eqn{w_{ij}} in the objective function.
+#' \eqn{O(n)}, is to draw unweighted samples and include \eqn{w_{ij}} in the objective function. In addition, the
+#' alternative probabalistic function used when \eqn{\alpha == 0} tends to overflow unless edge weights are used.
 #'
 #' Note that the input matrix should be symmetric.  If any columns in the matrix are empty, the function will fail.
 #'
@@ -29,7 +33,7 @@
 #' @param alpha Hyperparameter used in the default distance function, \eqn{1 / (1 + \alpha \dot ||y_i - y_j||^2)}.  If \code{alpha} is 0, the alternative distance
 #' function \eqn{1 / 1 + exp(||y_i - y_j||^2)} is used instead.  These functions relate the distance between points in the low-dimensional projection to the likelihood
 #' that they two points are nearest neighbors.
-#' @param weight_pos_samples Whether to sample positive edges according to their edge weights (the default) or take the
+#' @param weight_pos_samples Whether to sample positive edges according to their edge weights (the default, unless alpha == 0) or take the
 #' weights into account when calculating gradient.  See also the Details section.
 #' @param rho Initial learning rate.
 #' @param min_rho Final learning rate.
@@ -52,15 +56,15 @@ projectKNNs <- function(wij, # symmetric sparse matrix
                         dim = 2, # dimension of the projection space
                         sgd_batches = (length(wij@p) -1) * 20000,
                         M = 5,
-                        weight_pos_samples = TRUE,
+                        weight_pos_samples = if (alpha == 0) {FALSE} else {TRUE},
                         gamma = 7,
-                        alpha = 2,
+                        alpha = 1,
                         rho = 1,
                         coords = NULL,
-                        min_rho = 0.1,
+                        min_rho = 0,
                         verbose = TRUE) {
   N <-  (length(wij@p) -1)
-  js <- rep(0:(N-1), diff(wij@p))
+  js <- rep(0:(N - 1), diff(wij@p))
   is <- wij@i
 
   ##############################################