+++ R/projectionTreeSearch.R
@@ -7,7 +7,6 @@
 #' distinct partitionable clusters, try increasing the \code{tree_threshold} to increase the number
 #' of returned neighbors.
 #'
-#'
 #' @param x A matrix.
 #' @param K How many nearest neighbors to seek for each node.
 #' @param n_trees The number of trees to build.
@@ -15,6 +14,7 @@
 #' using a value equivalent to the number of features in the input set.
 #' @param max_iter Number of iterations in the neighborhood exploration phase.
 #' @param max_depth The maximum level of recursion.
+#' @param distance_method One of "Euclidean" or "Cosine."
 #' @param verbose Whether to print verbose logging using the \code{progress} package.
 #'
 #' @return A [K, N] matrix of the approximate K nearest neighbors for each vertex.
@@ -23,21 +23,21 @@
 #' @examples
 #'
 randomProjectionTreeSearch <- function(x,
-                                       K = 5, #
+                                       K = 5,
                                        n_trees = 2,
                                        tree_threshold =  max(10, nrow(x)),
                                        max_iter = 2,
                                        max_depth = 32,
+                                       distance_method = "Euclidean",
                                        verbose= TRUE) {
-  N <- nrow(x)
-
-  if (verbose) cat("Searching for neighbors.\n")
+   if (verbose) cat("Searching for neighbors.\n")
 
   knns <- searchTrees(threshold = tree_threshold,
                       n_trees = n_trees,
                       K = K, max_recursion_degree = max_depth,
                       maxIter = max_iter,
                       data = x,
+                      distance_method,
                       verbose = verbose)
 
   if (sum(colSums(knns != -1) == 0) + sum(is.na(knns)) + sum(is.nan(knns)) > 0)