+++ src/helpers.hpp
@@ -1,0 +1,9 @@
+#ifndef LARGEVISHELPERS
+#define LARGEVISHELPERS
+using namespace Rcpp;
+
+double dist(const arma::vec& i, const arma::vec& j);
+
+double cosDist(const arma::vec& i, const arma::vec& j);
+
+#endif