+++ src/largeVis.cpp
@@ -12,12 +12,9 @@ using namespace Rcpp;
 using namespace std;
 
 // The Euclidean distance between two vectors
-inline double dist(NumericVector i, NumericVector j) {
-  return sum(pow(i - j, 2));
-}
 
 inline double dist(arma::vec i, arma::vec j) {
-  return sum(pow(i - j, 2));
+  return sum(square(i - j));
 }
 
 /*
@@ -53,22 +50,21 @@ void checkGrad(const arma::vec& x,
 
 // [[Rcpp::export]]
 arma::mat sgd(arma::mat coords,
-         const NumericVector is, // vary randomly
-         const NumericVector js, // ordered
-         const NumericVector ps, // N+1 length vector of indices to start of each row j in vector is
-         const NumericVector ws, // w{ij}
-         const double gamma,
-         const double rho,
-         const double minRho,
-         const bool useWeights,
-         const int nBatches,
-         const int M,
-         const double alpha,
-         bool verbose) {
+              const NumericVector is, // vary randomly
+              const NumericVector js, // ordered
+              const NumericVector ps, // N+1 length vector of indices to start of each row j in vector is
+              const NumericVector ws, // w{ij}
+              const double gamma,
+              const double rho,
+              const double minRho,
+              const bool useWeights,
+              const int nBatches,
+              const int M,
+              const double alpha,
+              bool verbose) {
 
   Progress progress(nBatches, verbose);
 
-  const int updateRate = 1e6;
   const int D = coords.n_rows;
   const int N = ps.size() - 1;
   const int E = ws.length();
@@ -91,14 +87,11 @@ arma::mat sgd(arma::mat coords,
       positiveEdgeWeights[idx] = positiveEdgeWeights[idx - 1] + (ws[idx] / posScale);
   }
 
-  Rcpp::List progressParams = Rcpp::List::create();
-  double trailingLoss = 0;
-
   const int posSampleLength = min(1000000, nBatches);
   arma::vec positiveSamples = arma::randu<arma::vec>(posSampleLength);
 
   // Iterate through the edges in the positiveEdges vector
-#pragma omp parallel for shared(coords, positiveSamples, trailingLoss) schedule(static)
+#pragma omp parallel for shared(coords, positiveSamples) schedule(static)
   for (int eIdx=0; eIdx < nBatches; eIdx++) {
     if (progress.increment()) {
       const double posTarget = *(positiveSamples.begin() + (eIdx % posSampleLength));
@@ -119,28 +112,28 @@ arma::mat sgd(arma::mat coords,
 
       //if ((arma::randn<arma::vec>(1))[0] < 0) swap(i, j);
 
-      // const arma::vec y_i = coords.row(i);
-      // const arma::vec y_j = coords.row(j);
-      arma::vec y_i = arma::vec(D);
-      arma::vec y_j = arma::vec(D);
-      for (int idx = 0; idx < D; idx++) {
-        y_i[idx] = coords(idx,i);
-        y_j[idx] = coords(idx,j);
-      }
+      const arma::vec y_i = coords.col(i);
+      const arma::vec y_j = coords.col(j);
+      // arma::vec y_i = arma::vec(D);
+      // arma::vec y_j = arma::vec(D);
+      // for (int idx = 0; idx < D; idx++) {
+      //   y_i[idx] = coords(idx,i);
+      //   y_j[idx] = coords(idx,j);
+      // }
 
       // wij
       const double w = (useWeights) ? ws[e_ij] : 1;
 
-      const double dist_ij = (alpha == 0) ? std::min(8.0, sqrt(dist(y_i, y_j))) : sqrt(dist(y_i, y_j));
+      const double dist_ij = dist(y_i, y_j);
 
       const arma::vec d_dist_ij = (y_i - y_j) / sqrt(dist_ij);
       double p_ij;
-      if (alpha == 0)   p_ij =   1 / (1 +     exp( pow(dist_ij,2)));
-      else              p_ij =   1 / (1 + (alpha * pow(dist_ij,2)));
+      if (alpha == 0)   p_ij =   1 / (1 +      exp(dist_ij));
+      else              p_ij =   1 / (1 + (alpha * dist_ij));
 
       arma::vec d_p_ij;
-      if (alpha == 0) d_p_ij =  d_dist_ij * -2 * dist_ij * exp(pow(dist_ij,2)) / pow(1 + exp(pow(dist_ij,2)), 2);
-      else            d_p_ij =  d_dist_ij * -2 * dist_ij * alpha               / pow(1 +    (pow(dist_ij,2) * alpha),2);
+      if (alpha == 0) d_p_ij =  d_dist_ij * -2 * dist_ij * exp(dist_ij) / pow(1 + exp(dist_ij), 2);
+      else            d_p_ij =  d_dist_ij * -2 * dist_ij * alpha        / pow(1 +    (dist_ij * alpha),2);
 
       //double o = log(p_ij);
       const arma::vec d_j = (1 / p_ij) * d_p_ij;
@@ -162,26 +155,27 @@ arma::mat sgd(arma::mat coords,
                                std::upper_bound(negativeSampleWeights.begin(),
                                                 negativeSampleWeights.end(),
                                                 target)
-                                 );
+        );
 
         if (k == i ||
             k == j ||
             sum(searchVector == k) > 0) continue;
-        arma::vec y_k = arma::vec(D);
-        for (int idx = 0; idx < D; idx++) y_k[idx] = coords(idx,k);
+        const arma::vec y_k = coords.col(k);
+        // arma::vec y_k = arma::vec(D);
+        // for (int idx = 0; idx < D; idx++) y_k[idx] = coords(idx,k);
 
-        const double dist_ik = (alpha == 0) ? std::min(8.0, sqrt(dist(y_i, y_k))) : sqrt(dist(y_i, y_k));
+        const double dist_ik = dist(y_i, y_k);
         if (dist_ik == 0) continue; // Duplicates
 
         const arma::vec d_dist_ik = (y_i - y_k) / sqrt(dist_ik);
 
         double p_ik;
-        if (alpha == 0) p_ik  =  1 - (1 / (1 +     exp( pow(dist_ik,2))));
-        else            p_ik  =  1 - (1 / (1 + (alpha * pow(dist_ik,2))));
+        if (alpha == 0) p_ik  =  1 - (1 / (1 +      exp(dist_ik)));
+        else            p_ik  =  1 - (1 / (1 + (alpha * dist_ik)));
 
         arma::vec d_p_ik;
-        if (alpha == 0) d_p_ik =  d_dist_ik * 2 * dist_ik * exp(pow(dist_ik,2)) / pow(1 +      exp(pow(dist_ik,2)),2);
-        else            d_p_ik =  d_dist_ik * 2 * dist_ik * alpha               / pow(1 + (alpha * pow(dist_ik,2)),2);
+        if (alpha == 0) d_p_ik =  d_dist_ik * 2 * dist_ik * exp(dist_ik) / pow(1 +      exp(dist_ik),2);
+        else            d_p_ik =  d_dist_ik * 2 * dist_ik * alpha        / pow(1 + (alpha * dist_ik),2);
         //o += (gamma * log(p_ik));
 
         const arma::vec d_k = (gamma / p_ik) * d_p_ik;