+++ NEWS.md
@@ -1,9 +1,34 @@
-# largeVis 0.1.0
+
+
+
+### largeVis 0.1.4
+
+* Added option of Euclidean or Cosine distance. 
+* Now using the median in random projection trees, to make splits more even. This should eliminate the need for the
+max_depth parameter. 
+* The alternative (alpha == 0) probabilistic function now works. 
+* Benchmarks
+* Vignette
+* Vastly improved multi-threading performance
+
+### largeVis 0.1.3
+
+* Rewrote neighbor-search code to improve memory performance. 
+
+* Performance of wji calculation improved by an order of magnitude.
+
+### largeVis 0.1.2
+
+* Added visualization function.
+
+### largeVis 0.1.1
+
+* Many changes because of OpenMP compatibility issues. 
+
+* Moved from Rcpp to RcppArmadillo objects, etc. 
+
+### largeVis 0.1.0
+
+* Initial development releases.  Focused on correctness, performance, testing against larger datasets.
 
 * Added a `NEWS.md` file to track changes to the package.
-
-0.1.3 - Rewrote neighbor-search code to improve memory performance. Performance of wji calculation improved by an order of magnitude.
-0.1.2 - Added visualization function.
-0.1.1 - Many changes because of OpenMP compatibility issues. Moved from Rcpp to arma:: objects, etc. 
-0.1.0 - Initial development releases.  Focused on correctness, performance, testing against larger datasets.
-